+++ src/transformer/common/import-analyzer.ts
@@ -1,19 +1,17 @@
-import { TypeImport } from "./type-import";
-import { Visit, VisitorBase } from "./visitor-base";
+import { Visit } from "./visitor-base";
 import * as ts from 'typescript';
+import { RttiVisitor } from "../rtti-visitor-base";
+import { RttiContext } from "../rtti-context";
 
-export class ImportAnalyzer extends VisitorBase {
-    map = new Map<string,TypeImport>();
-
-    static analyze(file : ts.SourceFile, context : ts.TransformationContext): Map<string,TypeImport> {
-        let analyzer = new ImportAnalyzer(context);
+export class ImportAnalyzer extends RttiVisitor {
+    static analyze(file : ts.SourceFile, ctx : RttiContext) {
+        let analyzer = new ImportAnalyzer(ctx);
         try {
             analyzer.visitEachChild(file);
         } catch (e) {
             console.error(`RTTI: While analyzing imports for file '${file.fileName}': ${e.message}`);
             throw e;
         }
-        return analyzer.map;
     }
 
     @Visit(ts.SyntaxKind.ImportDeclaration)
@@ -24,7 +22,7 @@ export class ImportAnalyzer extends VisitorBase {
             if (!bindings) {
                 let name = decl.importClause.name.text;
 
-                this.map.set(name, {
+                this.ctx.importMap.set(name, {
                     name,
                     localName: name,
                     refName: name,
@@ -36,8 +34,9 @@ export class ImportAnalyzer extends VisitorBase {
             } else if (bindings) {
                 if (ts.isNamedImports(bindings)) {
                     for (let binding of bindings.elements) {
-                        this.map.set(binding.name.text, {
+                        this.ctx.importMap.set(binding.name.text, {
                             name: binding.name.text,
+                            symbol: this.checker.getSymbolAtLocation(binding.name),
                             localName: `${binding.propertyName?.text ?? binding.name.text}Φ`,
                             refName: binding.name.text,
                             modulePath: (<ts.StringLiteral>decl.moduleSpecifier).text,
@@ -48,7 +47,7 @@ export class ImportAnalyzer extends VisitorBase {
 
                         let nameAsInterface = `IΦ${binding.name.text}`;
 
-                        this.map.set(nameAsInterface, {
+                        this.ctx.importMap.set(nameAsInterface, {
                             name: nameAsInterface,
                             localName: nameAsInterface,
                             refName: nameAsInterface,
@@ -59,7 +58,7 @@ export class ImportAnalyzer extends VisitorBase {
                         });
                     }
                 } else if (ts.isNamespaceImport(bindings)) {
-                    this.map.set(bindings.name.text, {
+                    this.ctx.importMap.set(bindings.name.text, {
                         name: bindings.name.text,
                         localName: `${bindings.name.text}Φ`,
                         modulePath: (<ts.StringLiteral>decl.moduleSpecifier).text,