+++ src/transformer/common/type-import.ts
@@ -2,6 +2,7 @@ import * as ts from 'typescript';
 
 export interface TypeImport {
     importDeclaration : ts.ImportDeclaration;
+    symbol? : ts.Symbol;
     refName : string;
     modulePath : string;
     isNamespace : boolean;